+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This changelog is initialized in release 1.0.0
 
 ## [Unreleased]
 
+## [v2.5.0] - 2022-08-18
+
 ### Added
 * WhereIn method to repository
 * WhereNotIn method to repository
@@ -50,7 +52,8 @@ This changelog is initialized in release 1.0.0
 * Columns argument to ModelRepositoryInterface::all to be compatible with the underlying Eloquent model
 * FQN options for contract and repository to ModelRepositoryMakeCommand
 
-[Unreleased]: https://github.com/wimski/laravel-model-repositories/compare/v2.4.0...master
+[Unreleased]: https://github.com/wimski/laravel-model-repositories/compare/v2.5.0...master
+[v2.5.0]: https://github.com/wimski/laravel-model-repositories/compare/v2.4.0...v2.5.0
 [v2.4.0]: https://github.com/wimski/laravel-model-repositories/compare/v2.3.0...v2.4.0
 [v2.3.0]: https://github.com/wimski/laravel-model-repositories/compare/v2.2.0...v2.3.0
 [v2.2.0]: https://github.com/wimski/laravel-model-repositories/compare/v2.1.1...v2.2.0